+++ src/low/easy/relic_bn_mod_low.c
@@ -38,7 +38,8 @@
 /* Public definitions                                                         */
 /*============================================================================*/
 
-void bn_modn_low(dig_t *c, const dig_t *a, int sa, const dig_t *m, int sm, dig_t u) {
+void bn_modn_low(dig_t *c, const dig_t *a, size_t sa, const dig_t *m, size_t sm,
+		dig_t u) {
 	int i, j;
 	dig_t t, r0, r1, r2;
 	dig_t *tmp, *tmpc;
