+++ src/low/easy/relic_fb_mul_low.c
@@ -149,7 +149,7 @@ void fb_muln_low(dig_t *c, const dig_t *a, const dig_t *b) {
 	}
 }
 
-void fb_muld_low(dig_t *c, const dig_t *a, const dig_t *b, int size) {
+void fb_muld_low(dig_t *c, const dig_t *a, const dig_t *b, size_t size) {
     dig_t *tt = RLC_ALLOCA(dig_t, 16 * (size + 1));
 	dig_t *t[16];
 	dig_t u, r0, r1, r2, r4, r8, *tmpc;
